+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/SubstrateExitHandlerFeature.java
@@ -24,8 +24,6 @@
  */
 package com.oracle.svm.core;
 
-import org.graalvm.nativeimage.ImageInfo;
-
 import com.oracle.svm.core.annotate.Alias;
 import com.oracle.svm.core.annotate.TargetClass;
 import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
@@ -36,7 +34,7 @@ import com.oracle.svm.core.jdk.RuntimeSupport;
 public class SubstrateExitHandlerFeature implements InternalFeature {
     @Override
     public void beforeAnalysis(BeforeAnalysisAccess access) {
-        if (!ImageInfo.isSharedLibrary() || SubstrateOptions.DeprecatedOptions.InstallExitHandlers.getValue() || SubstrateOptions.InstallJavaExitHandlersForSharedLibrary.getValue()) {
+        if (SubstrateOptions.InstallExitHandlers.getValue()) {
             RuntimeSupport.getRuntimeSupport().addStartupHook(new SubstrateExitHandlerStartupHook());
         }
     }
@@ -45,10 +43,8 @@ public class SubstrateExitHandlerFeature implements InternalFeature {
 final class SubstrateExitHandlerStartupHook implements RuntimeSupport.Hook {
     @Override
     public void execute(boolean isFirstIsolate) {
-        if (SubstrateOptions.EnableSignalHandling.getValue()) {
-            if (isFirstIsolate) {
-                Target_java_lang_Terminator.setup();
-            }
+        if (SubstrateOptions.EnableSignalHandling.getValue() && isFirstIsolate) {
+            Target_java_lang_Terminator.setup();
         }
     }
 }